[BFB]

* jgfouca/scream_downstream_2024_11_07: (182 commits)
  Add ghci-oci
  Workflows: reworked how to skip eamxx testing jobs
  Workflows: added more eamxx workflows
  Workflows: fix logic for eamxx jobs skipping
  Workflows: allow to generate ALL baselines at ones for eamxx
  EAMxx: Use the main branch of mam4xx.
  EAMxx: Add tracer_reader_utils.hpp to microphysics.hpp.
  EAMxx: Remove temporary CPP macro.
  EAMxx: Update description of dry aerosol particle diameters.
  EAMxx: Use a temporary branch for mam4xx to test changes in the photo table code.
  EAMxx: Move read utils files to readfiles folder.
  EAMxx: Refine linoz parameter handling and class member references
  Correcting data types for variables, refining comments, and updating YAML input configurations.
  Addresses review comments: adds doc strings, reverts scorpio file change and cleanup
  Adds emissions file for ne4pg2, removes qaerwat from the input.yaml for single process test
  Using mam4::utils::extract_stateq_from_prognostics to copy data from prog to state_q.
  Refactor: move allocation of views from the tracer data struct interface to helper code. Combine multiple NC file registrations into a single function call: setup_tracer_data.
  Using one data structure to store beginning, end, and output data sets.
  Retrieve dgncur_a, dgncur_awet, wetdens, and qaerwat from FM
  1. Use the Scorpio interface to get altitude from vertical emission files. I was getting zeros or junk values when I employed the AtmInput class. 2. Add altitude_int as a member of the Altitude class.
  ...

+++ components/eamxx/src/physics/rrtmgp/eamxx_rrtmgp_process_interface.cpp
@@ -616,9 +616,9 @@ void RRTMGPRadiation::initialize_impl(const RunType /* run_type */) {
   // for duration of simulation.
   m_orbital_year = m_params.get<Int>("orbital_year",-9999);
   // Get orbital parameters from yaml file
-  m_orbital_eccen = m_params.get<Int>("orbital_eccentricity",-9999);
-  m_orbital_obliq = m_params.get<Int>("orbital_obliquity"   ,-9999);
-  m_orbital_mvelp = m_params.get<Int>("orbital_mvelp"       ,-9999);
+  m_orbital_eccen = m_params.get<double>("orbital_eccentricity",-9999);
+  m_orbital_obliq = m_params.get<double>("orbital_obliquity"   ,-9999);
+  m_orbital_mvelp = m_params.get<double>("orbital_mvelp"       ,-9999);
 
   // Determine whether or not we are using a fixed solar zenith angle (positive value)
   m_fixed_solar_zenith_angle = m_params.get<double>("Fixed Solar Zenith Angle", -9999);
